Fix story mode layout and consolidate redesigned components
