+++ frontend/src/main.jsx
@@ -1,8 +1,7 @@
 
-import "./index.css";   // path is relative to the current file
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import './index.css'
+import './index.css';
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
 import App from './App.jsx'
 
 createRoot(document.getElementById('root')).render(
